+++ src/com/xinyiglass/springSample/service/LgInvService.java
@@ -19,7 +19,7 @@ public class LgInvService {
 	PagePub pagePub;
 	
 	@Transactional(propagation=Propagation.NOT_SUPPORTED,readOnly=true)
-	public String findForPage(int pageSize,int pageNo,boolean goLastPage,String orderby,Long organizationId,Long width,Long height,Long thickness,String coatingType,Long loginId) throws Exception{
+	public String findForPage(int pageSize,int pageNo,boolean goLastPage,String orderby,Long organizationId,Long width,Long height,Double thickness,String coatingType,Long loginId) throws Exception{
 		Map<String,Object> paramMap=new HashMap<String,Object>();
 		StringBuffer sqlBuf=new StringBuffer();
 		sqlBuf.append("SELECT * FROM XYG_ALB2B_LG_ONHAND_V WHERE 1=1");
